Allow readline support to be disabled when building pc-parse on Unix
This is useful when building Unix DLLs (eg libpatr.so) and ensures that the
built library isn't unduly affected by what is or isn't installed on the build
machine.

The real change is in configure.in. The other files are generated by autoconf,
and exhibit a lot of churn due to the newer version of autoconf that was used.

Reviewed by: MarkS