Eliminate unnecessary libraries in Linux pcpatr shared library
We want to keep the library as small as possible, and these other libraries
aren't needed for the library's client API (as used by FieldWorks, at least;
they are used by the standalone version of the parser program).

It's possible that the link line should have used @LIBS@ instead of $(LIBS)
all along, in which case this just finishes off what was attempted in Tom's
recent commit.

Reviewed by: MarkS